decouple ChangesetTest from FileProcessorTest
