+++ packages/EasyCodingStandard/packages/SniffRunner/tests/Application/FixerTest.php
@@ -2,10 +2,10 @@
 
 namespace Symplify\EasyCodingStandard\SniffRunner\Tests\Application;
 
+use PHP_CodeSniffer\Fixer;
 use Symplify\EasyCodingStandard\HttpKernel\EasyCodingStandardKernel;
 use Symplify\EasyCodingStandard\SniffRunner\File\File;
 use Symplify\EasyCodingStandard\SniffRunner\File\FileFactory;
-use Symplify\EasyCodingStandard\SniffRunner\Fixer\Fixer;
 use Symplify\PackageBuilder\FileSystem\SmartFileInfo;
 use Symplify\PackageBuilder\Tests\AbstractKernelTestCase;
